requirements.txt
asgiref==3.11.0
Django==5.2.8
django-extensions==4.1
django-silk==5.4.3
djangorestframework==3.16.1
gprof2dot==2025.4.14
pillow==12.0.0
sqlparse==0.5.4
tzdata==2025.2
